@@ -4,81 +4,47 @@
   "license": "GPL-3.0",
   "private": true,
   "scripts": {
-    "build": "npx gulp build",
-    "start:dev": "npm run build && concurrently --kill-others \"npx gulp watch\" \"nodemon --watch ./backend ./backend/server.js\" \"firebase serve --only hosting\"",
-    "start:dev:nodb": "npm run build && concurrently --kill-others \"npx gulp watch\" \"firebase serve --only hosting\"",
-    "start:dev:nofe": "nodemon --watch ./backend ./backend/server.js",
-    "deploy:live": "npm run build && firebase deploy -P live --only hosting"
+    "start:dev": "concurrently --kill-others \"cd ./frontend && npm run start:dev\" \"cd ./backend && npm run start:dev\"",
+    "start:dev:be": "cd backend && npm run start:dev",
+    "start:dev:fe": "cd frontend && npm run start:dev",
+    "deploy:live": "cd frontend && npm run deploy:live",
+    "install:all": "sh ./bin/install.sh",
+    "install:windows": ".\\bin\\install.cmd",
+    "lint": "eslint \"./backend/**/*.{ts,js}\" \"./frontend/src/scripts/**/*.{ts,js}\"",
+    "build:live": "cd ./frontend && npm run build:live",
+    "pretty": "prettier --check './backend/**/*.js' './frontend/src/**/*.{js,scss}' './frontend/static/**/*.{json,html}'",
+    "pr-check-lint-json": "cd frontend && npx gulp pr-check-lint-json",
+    "pr-check-quote-json": "cd frontend && npx gulp pr-check-quote-json",
+    "pr-check-language-json": "cd frontend && npx gulp pr-check-language-json",
+    "pr-check-other-json": "cd frontend && npx gulp pr-check-other-json",
+    "pr-check-scss": "cd frontend && npx gulp pr-check-scss",
+    "pr-check-ts": "cd frontend && npx gulp pr-check-ts",
+    "pr-check-build-be": "cd backend && npm run build"
   },
   "engines": {
-    "node": "16.13.2",
     "npm": "8.1.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.3",
-    "@babel/plugin-transform-runtime": "^7.12.1",
-    "@babel/preset-env": "^7.12.1",
-    "babel-eslint": "^10.1.0",
-    "babelify": "^10.0.0",
-    "browserify": "^17.0.0",
-    "concurrently": "^5.3.0",
-    "dart-sass": "^1.25.0",
-    "del": "^6.0.0",
-    "eslint": "^7.28.0",
-    "eslint-config-standard": "^16.0.3",
-    "eslint-formatter-github": "^1.0.11",
-    "eslint-plugin-import": "^2.23.4",
-    "eslint-plugin-json": "^2.1.2",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^5.1.0",
-    "eslint-plugin-require-path-exists": "^1.1.9",
-    "gulp": "^4.0.2",
-    "gulp-concat": "^2.6.1",
-    "gulp-eslint": "^6.0.0",
-    "gulp-eslint-new": "^1.3.0",
-    "gulp-sass": "^5.0.0",
-    "gulp-uglify": "^3.0.2",
-    "husky": "^4.3.0",
-    "prettier": "2.1.2",
-    "pretty-quick": "^3.1.0",
-    "through2": "^4.0.2",
-    "vinyl-buffer": "^1.0.1",
-    "vinyl-paths": "^3.0.1",
-    "vinyl-source-stream": "^2.0.0"
+    "@typescript-eslint/eslint-plugin": "^5.11.0",
+    "@typescript-eslint/parser": "^5.11.0",
+    "concurrently": "5.3.0",
+    "eslint": "8.8.0",
+    "eslint-config-standard": "16.0.3",
+    "eslint-formatter-github": "1.0.11",
+    "eslint-plugin-import": "2.23.4",
+    "eslint-plugin-json": "2.1.2",
+    "eslint-plugin-node": "11.1.0",
+    "eslint-plugin-promise": "5.1.0",
+    "eslint-plugin-require-path-exists": "1.1.9",
+    "husky": "4.3.0",
+    "jsonschema": "1.4.0",
+    "prettier": "2.5.1",
+    "pretty-quick": "3.1.0",
+    "typescript": "^4.5.5"
   },
   "husky": {
     "hooks": {
-      "pre-commit": "pretty-quick --staged"
+      "pre-commit": "pretty-quick --staged && npm run lint"
     }
-  },
-  "dependencies": {
-    "@babel/runtime": "^7.12.5",
-    "axios": "^0.21.2",
-    "chart.js": "^2.9.4",
-    "chartjs-plugin-annotation": "^0.5.7",
-    "chartjs-plugin-trendline": "^0.2.2",
-    "cors": "^2.8.5",
-    "cron": "^1.8.2",
-    "dom-to-image": "^2.6.0",
-    "dotenv": "^10.0.0",
-    "express": "^4.17.1",
-    "express-rate-limit": "^5.3.0",
-    "firebase-admin": "^9.11.0",
-    "gulp-replace": "^1.1.3",
-    "helmet": "^4.6.0",
-    "howler": "^2.2.1",
-    "joi": "^17.6.0",
-    "lodash": "4.17.21",
-    "moment-timezone": "^0.5.33",
-    "mongodb": "^3.6.9",
-    "node-fetch": "^2.6.7",
-    "node-object-hash": "^2.3.10",
-    "nodemon": "^2.0.7",
-    "path": "^0.12.7",
-    "simple-git": "^2.45.1",
-    "string-similarity": "^4.0.4",
-    "tinycolor2": "^1.4.2",
-    "ua-parser-js": "^0.7.28",
-    "uuid": "^8.3.2"
   }
 }
