@@ -1,10 +1,10 @@
-const MonkeyError = require("../handlers/error");
-const { mongoDB } = require("../init/mongodb");
-const { ObjectID } = require("mongodb");
+import MonkeyError from "../utils/error";
+import db from "../init/db";
+import { ObjectId } from "mongodb";
 
 class PresetDAO {
   static async getPresets(uid) {
-    const preset = await mongoDB()
+    const preset = await db
       .collection("presets")
       .find({ uid })
       .sort({ timestamp: -1 })
@@ -13,9 +13,9 @@ class PresetDAO {
   }
 
   static async addPreset(uid, name, config) {
-    const count = await mongoDB().collection("presets").find({ uid }).count();
+    const count = await db.collection("presets").find({ uid }).count();
     if (count >= 10) throw new MonkeyError(409, "Too many presets");
-    let preset = await mongoDB()
+    let preset = await db
       .collection("presets")
       .insertOne({ uid, name, config });
     return {
@@ -25,30 +25,30 @@ class PresetDAO {
 
   static async editPreset(uid, _id, name, config) {
     console.log(_id);
-    const preset = await mongoDB()
+    const preset = await db
       .collection("presets")
-      .findOne({ uid, _id: ObjectID(_id) });
+      .findOne({ uid, _id: new ObjectId(_id) });
     if (!preset) throw new MonkeyError(404, "Preset not found");
     if (config) {
-      return await mongoDB()
+      return await db
         .collection("presets")
-        .updateOne({ uid, _id: ObjectID(_id) }, { $set: { name, config } });
+        .updateOne({ uid, _id: new ObjectId(_id) }, { $set: { name, config } });
     } else {
-      return await mongoDB()
+      return await db
         .collection("presets")
-        .updateOne({ uid, _id: ObjectID(_id) }, { $set: { name } });
+        .updateOne({ uid, _id: new ObjectId(_id) }, { $set: { name } });
     }
   }
 
   static async removePreset(uid, _id) {
-    const preset = await mongoDB()
+    const preset = await db
       .collection("presets")
-      .findOne({ uid, _id: ObjectID(_id) });
+      .findOne({ uid, _id: new ObjectId(_id) });
     if (!preset) throw new MonkeyError(404, "Preset not found");
-    return await mongoDB()
+    return await db
       .collection("presets")
-      .deleteOne({ uid, _id: ObjectID(_id) });
+      .deleteOne({ uid, _id: new ObjectId(_id) });
   }
 }
 
-module.exports = PresetDAO;
+export default PresetDAO;
